FocusBreakBot/requirements.txt
requests
apscheduler
python-telegram-bot
python-dotenv
python-telegram-bot==20.7